Extract module data store to its own file
